requirements.txt
pydub==0.25.1
requests~=2.25.1
RPi.GPIO==0.7.0